app/requirements.txt
asgiref==3.8.1
Flask==3.1.0
Flask-Cors==5.0.0
python-telegram-bot==21.7
fastapi==0.115.2